moving auth out
