+++ r2r/main/api/routes/retrieval.py
@@ -15,8 +15,7 @@ class RetrievalRouter(BaseRouter):
     def __init__(
         self, engine: R2REngine, auth_handler: Optional[AuthHandler] = None
     ):
-        super().__init__(engine)
-        self.auth_handler = auth_handler
+        super().__init__(engine, auth_handler)
         self.setup_routes()
 
     def setup_routes(self):
